+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- feat(cli): add output handler support for JavaScript and Python to process evaluation results programmatically (#4760)
+
+### Changed
+
 - chore: Add visiblity button for PFX passphrase (#6258)
 
 ## [0.119.8] - 2025-11-18